[Minuit2] Merge parameter transformation classes into one file
These classes are very small, it's not necessary to split things up in
different files at that granularity.
