+++ math/minuit2/inc/Minuit2/NumericalDerivator.h
@@ -21,9 +21,7 @@
 #include <Math/IFunctionfwd.h>
 
 #include "Fit/ParameterSettings.h"
-#include "Minuit2/SinParameterTransformation.h"
-#include "Minuit2/SqrtUpParameterTransformation.h"
-#include "Minuit2/SqrtLowParameterTransformation.h"
+#include "Minuit2/MnParameterTransformation.h"
 #include "Minuit2/MnMachinePrecision.h"
 
 #include <ROOT/RSpan.hxx>
