+++ math/minuit2/src/CMakeLists.txt
@@ -64,6 +64,7 @@ set(MINUIT2_HEADERS
     MnParabola.h
     MnParabolaFactory.h
     MnParameterScan.h
+    MnParameterTransformation.h
     MnPlot.h
     MnPoint.h
     MnPosDef.h
@@ -87,9 +88,6 @@ set(MINUIT2_HEADERS
     SimplexMinimizer.h
     SimplexParameters.h
     SimplexSeedGenerator.h
-    SinParameterTransformation.h
-    SqrtLowParameterTransformation.h
-    SqrtUpParameterTransformation.h
     VariableMetricBuilder.h
     VariableMetricEDMEstimator.h
     VariableMetricMinimizer.h
@@ -122,11 +120,12 @@ set(MINUIT2_SOURCES
     MnGlobalCorrelationCoeff.cxx
     MnHesse.cxx
     MnLineSearch.cxx
-    MnMatrix.cxx
     MnMachinePrecision.cxx
+    MnMatrix.cxx
     MnMinos.cxx
     MnParabolaFactory.cxx
     MnParameterScan.cxx
+    MnParameterTransformation.cxx
     MnPlot.cxx
     MnPosDef.cxx
     MnPrint.cxx
@@ -146,9 +145,6 @@ set(MINUIT2_SOURCES
     SimplexBuilder.cxx
     SimplexParameters.cxx
     SimplexSeedGenerator.cxx
-    SinParameterTransformation.cxx
-    SqrtLowParameterTransformation.cxx
-    SqrtUpParameterTransformation.cxx
     VariableMetricBuilder.cxx
     VariableMetricEDMEstimator.cxx
     mnbins.cxx
